refactor client to use crtp and support http 1.1
Refactor the http client to use crtp to reduce implementation
duplication. The http and https classes now inherit from a common base
http class templated with the derived class.

The http client now supports sending multiple requests over the same tcp
socket by reading request objects from a queue. On the final request,
the client will automatically send the connection close header.
Successful requests are removed from the queue. If an error is
encountered, the client will stop and return with the unfinished request
objects still in the queue. The remaining requests in the queue can be
sent by invoking the connect function again.

The http client now has a max timeout variable, allowing a maximum time
in milliseconds to be specified for the client to run.